backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
anthropic==0.39.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
playwright==1.49.1

# After installing requirements, run:
# playwright install chromium